+++ src/preload/index.ts
@@ -735,7 +735,8 @@ const api: ApplicationAPI = {
 
   getCommands: (baseDir) => ipcRenderer.invoke('get-commands', baseDir),
   getCustomModes: (baseDir) => ipcRenderer.invoke('get-custom-modes', baseDir),
-  runCustomCommand: (baseDir, taskId, commandName, args, mode) => ipcRenderer.invoke('run-custom-command', baseDir, taskId, commandName, args, mode),
+  runCustomCommand: (baseDir, taskId, commandName, args, mode, images) =>
+    ipcRenderer.invoke('run-custom-command', baseDir, taskId, commandName, args, mode, images),
 
   // Terminal operations
   isTerminalSupported: () => true,